Added utility to stringify kernel sources.
This is a C-only utility.  I'm working on adding Fortran support.  However, this is slightly more complicated.  Also, I would like to begin using this method of kernel definition in the test suite, but I'll have to alter the generation script to add a BUILT_SOURCES target.  Overall, this approach will get rid of the error prone environment variable search path method that is currently in use.