Added scoped event list behavior.
Extended ocl_initialize_event_wait_list to take an array of cl_event to
use as opposed to dynamically allocating space.  This obviates the call
to ocl_clear_event_wait_list.  If a local events array is passed, it
simply goes out of scope at the end of the translation unit.