Fixed string functions.
Several functions can be undefined when using the default flags for this
project.  This check-in provides alternate implementations.  in
particular, I have gotten rid of _GNU_SOURCE.  This appears to cause
problems for strdup, strndup and strsep.  These are now defined locallly
when they are not detected in the environment.