Created new sandbox area.
The new sandbox directory is for experiments that are not to be part of
a distribution (still need to figure out how to exclude the files from a
dist).  Also updated kernel_info to be kernel_work_group_info.  Need to
still do some revamping of the kernel hints to take into account the
amount of local and private memory are used by a kernel.  This is not
necessarily straight forward.  Still need to consider how it
should/could be done.