Added counter to help in initialization.
When initializing more than one device of a certain type, you may need
to know how many have been created already.  This commit adds counters
for each device type, so that this information is available.  The
generic initialization method now uses this to increment the device id
of new devices.